+++ movieshopspa/src/app/account/account-routing.module.ts
@@ -1,7 +1,9 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 
-const routes: Routes = [];
+const routes: Routes = [
+  
+];
 
 @NgModule({
   imports: [RouterModule.forChild(routes)],
