+++ movieshopspa/src/app/movies/movie-details/movie-details.component.html
@@ -1,1 +1,94 @@
-<p>movie-details works!</p>
+
+
+    <div class="row" style="background-image: url({{movie.posterUrl}})">
+        <div class="col-md-3 style=width: 10rem">           
+                <img src="{{movie.posterUrl}}" alt="{{movie.title}}" />
+        </div>         
+        <div class="col-6">
+                <ul class="list-unstyled">
+                    <li class="list-group-item align-content-center"><h2>{{movie.title}}</h2></li>
+                    <li class="list-group-item align-content-center" *ngFor="Let genre of genre">
+                        <span class="badge rounded-pill bg-secondary">{{movie.genres}}</span>
+                    </li>
+                    <!--li><span class="badge" [ngClass]="{'bg-success': {{movie.rating}} >= 9">{{movie.rating}}</span></-li-->
+                    <li class="list-group-item align-content-center">{{movie.overview}}</li>
+                </ul>
+            </div>
+            <div class="col-md-2 gap-2">
+                <form>
+                    <input type=submit value ="Review Movie" class="btn btn-primary"> 
+                </form>
+                <form asp-controller="User" asp-action="BuyMovies" method="post">
+                    <button type="button" class="btn btn-primary" data-bs-toggle="modal" data-bs-target="#staticBackdrop">${{movie.price | currency}}</button>
+                    
+                    <!-- Buy Button Popup Model -->
+                    <div class="modal fade" id="staticBackdrop" data-bs-backdrop="static" data-bs-keyboard="false" tabindex="-1" aria-labelledby="staticBackdropLabel" aria-hidden="true">
+                        <div class="modal-dialog">
+                            <div class="modal-content">
+                                <div class="modal-header">
+                                    <h6 class="modal-title" id="staticBackdropLabel">{{movie.title}}</h6>
+                                        <input type="hidden" value="{{movie.id}}" name="{{movie.id}}">
+                                        <button type="button" class="btn-close" data-bs-dismiss="modal" aria-label="Close"></button>
+                                </div>
+                                <div class="modal-body">
+                                    <div class="col-3" style="width: 10rem;">
+                                        <img src="{{movie.posterUrl}}" class="card-img-top" alt={{movie.title}}>              
+                                    </div>
+                                </div>
+                                <div class="modal-footer">
+                                    <button type="button" class="btn btn-secondary" data-bs-dismiss="modal">Close</button>
+                                    <input type=submit value = "Buy $@Model.Price"  name="btnBuyMovies" class="btn btn-primary">
+                                </div>
+                            </div>
+                        </div>
+                    </div>  
+                    <!--div class ="row pt-1">
+                        <button type="button" class="btn-outline-success" data-bs-toggle="modal" data-bs-target="#staticBackdrop">WatchMovie</button>
+                    </!--div-->                
+                </form>
+        </div>
+            
+
+        </div>
+
+
+
+<!-- using ngClass directive to add/remove css classes based on condition -->
+<div class="row">
+    <div class="col-3 col-xl-3 col-sm-2 col-lg-3">
+      <h4>MOVIE FACTS</h4>
+       <ul class="list-unstyled">
+          <li class="list-group-item"><a>Release Date : </a><span class="badge rounded-pill bg-secondary">{{movie.releaseDate}}</span></li>
+          <li class="list-group-item"><a>Run Time : </a><span class="badge rounded-pill bg-secondary">{{movie.runTime}}</span></li>
+          <li class="list-group-item"><a>Box Office :  </a><span class="badge rounded-pill bg-secondary">{{movie.revenue | currency}}</span></li>
+          <li class="list-group-item"><a>Budget : </a><span class="badge rounded-pill bg-secondary">{{movie.budget | currency}}</span></li>  
+        </ul>
+        <div>
+        <h3>Trailers</h3>    
+        <!--ul class="list-unstyled">
+            @foreach (var Model in Model.Trailers)
+                {    
+                    <li><a href="@Model.TrailerUrl">@Model.Name &nbsp</a></li>
+                }
+        </!--ul--> 
+        </div>
+    </div>
+    <div class="col-6">
+        <h5>CAST</h5>
+      <!--ul class="list-unstyled">
+             @foreach (var Model in Model.Casts)
+                {    <li><img src="@Model.ProfilePath" alt="@Model.Name" width="40" height="40">
+                    <a href="@Model.ProfilePath">@Model.Name &nbsp</a> <a>@Model.Character</a></li>
+                }
+        </!--ul-->
+    </div>
+    <div class="col-3">  
+        
+    </div>
+  </div>
+<!--h4>
+    Movie Rating :<span class="badge" [ngClass]="{'bg-success': movie.rating >= 9, 'bg-danger' : movie.rating==0}">
+        {{movie.rating}} </span>
+</!--h4-->
+
+<!-- Using Pipes -->
