+++ movieshopspa/src/app/movies/movies.component.html
@@ -1,2 +1,2 @@
-<p>movies works!</p>
+
 <router-outlet></router-outlet>
