+++ movieshopspa/src/app/shared/models/Cast.ts
@@ -1,0 +1,7 @@
+
+export interface Cast {
+    id: number;
+    name: string;
+    profilePath: string;
+    character: string;
+}
