+++ movieshopspa/src/app/shared/models/Genre.ts
@@ -1,0 +1,5 @@
+
+export interface Genre {
+    id: number;
+    name: string;
+}
