+++ movieshopspa/src/app/shared/models/Trailer.ts
@@ -1,0 +1,6 @@
+
+export interface Trailer {
+    id: number;
+    trailerUrl: string;
+    name: string;
+}
