+++ movieshopspa/src/app/shared/models/movieDetails.ts
@@ -1,5 +1,9 @@
 // Generated by https://quicktype.io
 
+import { Cast } from "./Cast";
+import { Genre } from "./Genre";
+import { Trailer } from "./Trailer";
+
 export interface MovieDetails {
     id:          number;
     title:       string;
@@ -20,20 +24,4 @@ export interface MovieDetails {
     casts:       Cast[];
 }
 
-export interface Cast {
-    id:          number;
-    name:        string;
-    profilePath: string;
-    character:   string;
-}
 
-export interface Genre {
-    id:   number;
-    name: string;
-}
-
-export interface Trailer {
-    id:         number;
-    trailerUrl: string;
-    name:       string;
-}
